multi_env_mlops/requirements.txt
azure-ai-ml>=1.20
azure-identity>=1.17
pandas>=2.2
numpy>=2.0
